visualizations/requirements.txt
streamlit==1.42.2
pandas==2.2.3
plotly==6.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
statsmodels==0.14.4
watchdog==6.0.0